requirements.txt
svgwrite
Pillow